Added a LIDAR demo mode to LAMP.
The code works now.  As requested by Prof. Little, the light changes to
green when the LIDAR sensor detects something (close) under it, and is red
otherwise.

When the demo is lifted to the ceiling, the distance threshold (between
red and green) will need to be adjusted.

The code is a bit messy now; TODO: try refactoring the code.